# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_call_service.RejectCallStruct:
    properties:
      callCreator:
        $ref: '#/definitions/types.JID'
      callId:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct:
    properties:
      chat:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_chat_service.HistorySyncRequestStruct:
    properties:
      count:
        type: integer
      messageInfo:
        $ref: '#/definitions/types.MessageInfo'
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_community_service.AddParticipantStruct:
    properties:
      communityJid:
        type: string
      groupJid:
        items:
          type: string
        type: array
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_community_service.CreateCommunityStruct:
    properties:
      communityName:
        type: string
    type: object
...
  github_com_EvolutionAPI_evolution-go_pkg_group_service.CreateGroupStruct:
    properties:
      groupName:
        type: string
      participants:
        items:
          type: string
        type: array
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_group_service.GetGroupInfoStruct:
    properties:
      groupJid:
        type: string
    type: object
...
  github_com_EvolutionAPI_evolution-go_pkg_group_service.JoinGroupStruct:
    properties:
      code:
        type: string
    type: object
...
  github_com_EvolutionAPI_evolution-go_pkg_group_service.SetGroupDescriptionStruct:
    properties:
      description:
        type: string
      groupJid:
        type: string
    type: object
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
  /call/reject:
    post:
      consumes:
      - application/json
      description: Reject call
      parameters:
      - description: Call data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_call_service.RejectCallStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Reject call
      tags:
      - Call
  /chat/archive:
    post:
      consumes:
      - application/json
      description: Archive a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Archive a chat
      tags:
      - Chat
  /chat/history-sync:
    post:
      consumes:
      - application/json
      description: HistorySyncRequest a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.HistorySyncRequestStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: HistorySyncRequest a chat
      tags:
      - Chat
  /chat/mute:
    post:
      consumes:
      - application/json
      description: Mute a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Mute a chat
      tags:
      - Chat
  /chat/pin:
    post:
      consumes:
      - application/json
      description: Pin a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Pin a chat
      tags:
      - Chat
  /chat/unarchive:
    post:
      consumes:
      - application/json
      description: Unarchive a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Unarchive a chat
      tags:
      - Chat
  /chat/unmute:
    post:
      consumes:
      - application/json
      description: Unmute a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Unmute a chat
      tags:
      - Chat
  /chat/unpin:
    post:
      consumes:
      - application/json
      description: Unpin a chat
      parameters:
      - description: Chat
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_chat_service.BodyStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Unpin a chat
      tags:
      - Chat
  /community/add:
    post:
      consumes:
      - application/json
      description: Add participant to community
      parameters:
      - description: Participant data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_community_service.AddParticipantStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Add participant to community
      tags:
      - Community
  /community/create:
    post:
      consumes:
      - application/json
      description: Create community
      parameters:
      - description: Community data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_community_service.CreateCommunityStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create community
      tags:
      - Community
  /community/remove:
    post:
      consumes:
      - application/json
      description: Remove participant from community
      parameters:
      - description: Participant data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_community_service.AddParticipantStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Remove participant from community
      tags:
      - Community
  /group/create:
    post:
      consumes:
      - application/json
      description: Create group
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.CreateGroupStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create group
      tags:
      - Group
  /group/description:
    post:
      consumes:
      - application/json
      description: Set group description
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.SetGroupDescriptionStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set group description
      tags:
      - Group
  /group/info:
    post:
      consumes:
      - application/json
      description: Get group info
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.GetGroupInfoStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get group info
      tags:
      - Group
  /group/invitelink:
    post:
      consumes:
      - application/json
      description: Get group invite link
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.GetGroupInviteLinkStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get group invite link
      tags:
      - Group
  /group/join:
    post:
      consumes:
      - application/json
      description: Join group link
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.JoinGroupStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Join group link
      tags:
      - Group
  /group/leave:
    post:
      consumes:
      - application/json
      description: Leave group
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.LeaveGroupStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Leave group
      tags:
      - Group
  /group/list:
    get:
      consumes:
      - application/json
      description: List groups
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List groups
      tags:
      - Group
  /group/myall:
    get:
      consumes:
      - application/json
      description: Get my groups
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get my groups
      tags:
      - Group
  /group/name:
    post:
      consumes:
      - application/json
      description: Set group name
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.SetGroupNameStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set group name
      tags:
      - Group
  /group/participant:
    post:
      consumes:
      - application/json
      description: Update participant
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.AddParticipantStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update participant
      tags:
      - Group
  /group/photo:
    post:
      consumes:
      - application/json
      description: Set group photo
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.SetGroupPhotoStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Set group photo
      tags:
      - Group
  /group/settings:
    post:
      consumes:
      - application/json
      description: Update group settings (announcement, not_announcement, locked,
        unlocked, approval_on, approval_off, admin_add, all_member_add)
      parameters:
      - description: Group data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_group_service.UpdateGroupSettingsStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update group settings
      tags:
      - Group
  /instance/{instanceId}/advanced-settings:
    get:
      description: Get advanced settings for a specific instance
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Advanced settings retrieved successfully
          schema:
            $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_model.AdvancedSettings'
        "400":
          description: Invalid instance ID
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Instance not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get advanced settings
      tags:
      - Instance
    put:
      consumes:
      - application/json
      description: Update advanced settings for a specific instance
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Advanced settings data
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_model.AdvancedSettings'
      produces:
      - application/json
      responses:
        "200":
          description: Advanced settings updated successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Instance not found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update advanced settings
      tags:
      - Instance
  /instance/all:
    get:
      consumes:
      - application/json
      description: Get all instances
      produces:
      - application/json
      responses:
        "200":
          description: All instances
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get all instances
      tags:
      - Instance
  /instance/connect:
    post:
      consumes:
      - application/json
      description: Connect to instance with the provided data
      parameters:
      - description: Instance data
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_instance_service.ConnectStruct'
      produces:
      - application/json
      responses:
        "200":
          description: Instance connected successfully
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Error on validation
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Connect to instance
      tags:
      - Instance
...
swagger: "2.0"